Extract to `common` prover and verifier.
