+++ prover/src/common.rs
@@ -1,0 +1,5 @@
+mod prover;
+mod verifier;
+
+pub use self::{prover::Prover, verifier::Verifier};
+pub use aggregator::ChunkHash;
