+++ prover/src/lib.rs
@@ -1,4 +1,5 @@
 pub mod aggregator;
+pub mod common;
 pub mod config;
 mod evm_verifier;
 pub mod io;
@@ -7,6 +8,7 @@ pub mod test_util;
 pub mod utils;
 pub mod zkevm;
 
+pub use common::ChunkHash;
 pub use evm_verifier::EvmVerifier;
 pub use proof::Proof;
